+++ CHANGELOG.md
@@ -4,10 +4,22 @@
 
 ### Added
 
+### Changed
+
+## 0.5.0
+
+### Added
+
 - Support local tiff files.
 - Support local zarr directories.
+- onClick callbacks for layers.
 
 ### Changed
+
+- Removed greedy matching of dimension order in Bio-Formats Zarr output. Just check if OME-Zarr.
+- Upgrade to deck.gl 8.3.0
+- Use WASM LZW decoder from `manzt/geotiff.js`
+- Bundle UMD build with deps.
 
 ## 0.4.2